refactoring(15): extract pure ancestor-chain helper from DetailPanel
Move the parent_id walk (the only module-pure unit in the DetailPanel
monolith) into Tree/detailPanelHelpers.ts as computeAncestorIds; the
container's ancestorIds useMemo now calls it. Behavior byte-identical.
